// frontend/src/app/swagger/model/meatPiecePart.ts
/**
 * OrderBackend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeatPiece } from './meatPiece';


export interface MeatPiecePart { 
    id?: number;
    weight?: number;
    notes?: string | null;
    meatPieceId?: number;
    meatPiece?: MeatPiece;
}