// frontend/src/app/swagger/model/subCategoryDto.ts
/**
 * OrderBackend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeatPieceDto } from './meatPieceDto';


export interface SubCategoryDto { 
    id?: number;
    name?: string | null;
    meatPieces?: Array<MeatPieceDto> | null;
}